Add apidoc for login Prevent data leak between users

+++ src/controllers/users.controller.ts
@@ -1,17 +1,25 @@
-import { NextFunction, Request, Response } from 'express';
+import {NextFunction, Request, Response} from 'express';
 import {CreatedUser, CreateUserDto} from '../dtos/users.dto';
-import { User } from '../interfaces/users.interface';
+import {User} from '../interfaces/users.interface';
 import UserService from '../services/users.service';
 import BlockchainService from "../services/blockchain.service";
+import {RequestWithUser} from "../interfaces/auth.interface";
+import HttpException from "../exceptions/HttpException";
 
 class UsersController {
   public userService = new UserService();
   public blockchainService = new BlockchainService();
 
-  public getUserById = async (req: Request, res: Response, next: NextFunction) => {
+  private static verifyUserHasAccess(id: number, user: User): void {
+    if (user.id !== id)
+      throw new HttpException(403, "You don't have access to this resource");
+  }
+
+  public getUserById = async (req: RequestWithUser, res: Response, next: NextFunction) => {
     const userId: number = Number(req.params.id);
 
     try {
+      UsersController.verifyUserHasAccess(userId, req.user);
       const findOneUserData: CreatedUser = await this.userService.findUserById(userId);
       res.status(200).json({ data: findOneUserData, message: 'findOne' });
     } catch (error) {
@@ -31,11 +39,11 @@ class UsersController {
     }
   }
 
-  public updateUser = async (req: Request, res: Response, next: NextFunction) => {
+  public updateUser = async (req: RequestWithUser, res: Response, next: NextFunction) => {
     const userId: number = Number(req.params.id);
     const userData: User = req.body;
-
     try {
+      UsersController.verifyUserHasAccess(userId, userData);
       const updateUserData: CreatedUser = await this.userService.updateUser(userId, userData);
       res.status(200).json({ data: updateUserData, message: 'updated' });
     } catch (error) {
@@ -43,10 +51,11 @@ class UsersController {
     }
   }
 
-  public deleteUser = async (req: Request, res: Response, next: NextFunction) => {
+  public deleteUser = async (req: RequestWithUser, res: Response, next: NextFunction) => {
     const userId: number = Number(req.params.id);
 
     try {
+      UsersController.verifyUserHasAccess(userId, req.user);
       await this.userService.deleteUserData(userId);
       res.status(200).json({ message: 'deleted' });
     } catch (error) {
